backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlmodel>=0.0.22
pydantic>=2.4.0
python-dotenv>=1.0.0

